// SDKs/twilio.com/twilio_taskrouter_v1/1.36.0/typescript/src/sdk/models/operations/fetchtaskqueuecumulativestatistics.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";
export const FETCHTASKQUEUECUMULATIVESTATISTICS_SERVERS = [
	"https://taskrouter.twilio.com",
];



export class FetchTaskQueueCumulativeStatisticsPathParams extends SpeakeasyBase {
  @Metadata({ data: "pathParam, style=simple;explode=false;name=TaskQueueSid" })
  taskQueueSid: string;

  @Metadata({ data: "pathParam, style=simple;explode=false;name=WorkspaceSid" })
  workspaceSid: string;
}


export class FetchTaskQueueCumulativeStatisticsQueryParams extends SpeakeasyBase {
  @Metadata({ data: "queryParam, style=form;explode=true;name=EndDate" })
  endDate?: Date;

  @Metadata({ data: "queryParam, style=form;explode=true;name=Minutes" })
  minutes?: number;

  @Metadata({ data: "queryParam, style=form;explode=true;name=SplitByWaitTime" })
  splitByWaitTime?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=StartDate" })
  startDate?: Date;

  @Metadata({ data: "queryParam, style=form;explode=true;name=TaskChannel" })
  taskChannel?: string;
}


export class FetchTaskQueueCumulativeStatisticsSecurity extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class FetchTaskQueueCumulativeStatisticsRequest extends SpeakeasyBase {
  @Metadata()
  serverUrl?: string;

  @Metadata()
  pathParams: FetchTaskQueueCumulativeStatisticsPathParams;

  @Metadata()
  queryParams: FetchTaskQueueCumulativeStatisticsQueryParams;

  @Metadata()
  security: FetchTaskQueueCumulativeStatisticsSecurity;
}


export class FetchTaskQueueCumulativeStatisticsResponse extends SpeakeasyBase {
  @Metadata()
  contentType: string;

  @Metadata()
  statusCode: number;

  @Metadata()
  taskrouterV1WorkspaceTaskQueueTaskQueueCumulativeStatistics?: shared.TaskrouterV1WorkspaceTaskQueueTaskQueueCumulativeStatistics;
}
